Move kompassiGame.ts
