+++ shared/typings/models/kompassiGame/kompassiGame.ts
